pyproject.toml
[project]
name = "ticintel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "backboard-sdk>=1.5.14",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.46.0",
]
